+++ mvvm-coroutines-clean-architecture/domain/src/main/java/br/com/mobiplus/gitclient/domain/model/GitRepoModel.kt
@@ -20,5 +20,5 @@ data class GitRepoModel(
     val watchers: Int? = null,
     val watchersCount: Int? = null,
     val stargazersCount: Int? = null,
-    var reliabilityFactor: Int? = null
+    var reliabilityFactor: Double? = null
 )
